fix(py): preserve native structure count widths
- Return usize for native residue, chain, and model counts.
- Avoid lossy conversion and the forbidden unwrap suggestion from rustc.
- Verify clippy, native tests, and the Python pipeline in a clean worktree.

+++ crates/molframe-py/src/bindings.rs
@@ -74,17 +74,17 @@ impl PyStructure {
     }
 
     #[getter]
-    fn residue_count(&self) -> u32 {
+    fn residue_count(&self) -> usize {
         self.inner.residue_count()
     }
 
     #[getter]
-    fn chain_count(&self) -> u32 {
+    fn chain_count(&self) -> usize {
         self.inner.chain_count()
     }
 
     #[getter]
-    fn model_count(&self) -> u32 {
+    fn model_count(&self) -> usize {
         self.inner.model_count()
     }
